Moving SoundSliderPanel to Controls as SoundSlider
