+++ Autobox.Desktop/Activities/Panels/TrackPlayerPanel.xaml
@@ -1,9 +1,10 @@
-﻿<UserControl x:Class="Autobox.Desktop.Activities.Panels.TrackPlayerPanel"
+﻿<UserControl
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:Autobox.Desktop.Activities.Panels"
+             xmlns:Controls="clr-namespace:Autobox.Desktop.Activities.Controls" x:Class="Autobox.Desktop.Activities.Panels.TrackPlayerPanel"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid Background="{StaticResource PrimaryColor}">
@@ -21,6 +22,7 @@
                 <ColumnDefinition Width="{StaticResource H1.GridLength}"/>
                 <ColumnDefinition Width="{StaticResource H3.GridLength}"/>
             </Grid.ColumnDefinitions>
+
             <Grid.RowDefinitions>
                 <RowDefinition Height="{StaticResource H3.GridLength}"/>
                 <RowDefinition/>
@@ -28,6 +30,9 @@
                 <RowDefinition Height="{StaticResource H1.GridLength}"/>
                 <RowDefinition Height="{StaticResource H3.GridLength}"/>
             </Grid.RowDefinitions>
+
+            <Controls:SoundSlider x:Name="SoundSlider" Grid.Column="3" Grid.Row="1" Grid.RowSpan="3" ValueChanged="SoundSlider_ValueChanged"/>
+
             <DockPanel Grid.Column="1" Grid.Row="3" LastChildFill="True" >
                 <Button x:Name="ShuffleButton" Style="{StaticResource IconButton}" Click="ShuffleButton_Click">
                     <Button.OpacityMask>
@@ -46,8 +51,8 @@
                 </Button>
                 <Separator Width="{StaticResource H3.Double}" Background="Transparent"/>
             </DockPanel>
+
             <local:ScreenPanel x:Name="ScreenPanel" Grid.Column="1" Grid.Row="1"/>
-            <local:SoundSliderPanel x:Name="SoundSliderPanel" Grid.Column="3" Grid.Row="1" Grid.RowSpan="3" ValueChanged="SoundSliderPanel_ValueChanged"/>
         </Grid>
     </Grid>
 </UserControl>
