+++ Autobox.Desktop/Activities/Panels/TrackPreviewPanel.xaml
@@ -1,9 +1,10 @@
-﻿<UserControl x:Class="Autobox.Desktop.Activities.Panels.TrackPreviewPanel"
+﻿<UserControl
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:Autobox.Desktop.Activities.Panels"
+             xmlns:Controls="clr-namespace:Autobox.Desktop.Activities.Controls" x:Class="Autobox.Desktop.Activities.Panels.TrackPreviewPanel"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
 
@@ -22,6 +23,7 @@
                 <ColumnDefinition Width="{StaticResource H1.GridLength}"/>
                 <ColumnDefinition Width="{StaticResource H3.GridLength}"/>
             </Grid.ColumnDefinitions>
+            
             <Grid.RowDefinitions>
                 <RowDefinition Height="{StaticResource H3.GridLength}"/>
                 <RowDefinition/>
@@ -29,6 +31,7 @@
                 <RowDefinition Height="{StaticResource H1.GridLength}"/>
                 <RowDefinition Height="{StaticResource H3.GridLength}"/>
             </Grid.RowDefinitions>
+            
             <Rectangle Grid.Column="1" Panel.ZIndex="-2" Fill="Black" Grid.Row="1" Stroke="Black"/>
             <Image x:Name="ThumbnailPlaceholder" Panel.ZIndex="-1" Grid.Column="1" Grid.Row="1"/>
             <MediaElement x:Name="MediaPlayer" Grid.Column="1" Grid.Row="1" LoadedBehavior="Manual"/>
@@ -45,7 +48,7 @@
                 </DockPanel>
             </DockPanel>
             <Separator Width="{StaticResource H3.Double}" Background="{x:Null}" />
-            <local:SoundSliderPanel x:Name="SoundSliderPanel" Grid.Column="3" Grid.Row="1" Grid.RowSpan="3" ValueChanged="SoundSliderPanel_ValueChanged"/>
+            <Controls:SoundSlider x:Name="SoundSlider" Grid.Column="3" Grid.Row="1" Grid.RowSpan="3" ValueChanged="SoundSlider_ValueChanged"/>
         </Grid>
     </Grid>
 </UserControl>
