+++ Autobox.Desktop/Autobox.Desktop.csproj
@@ -63,6 +63,12 @@
     <Compile Include="Activities\ActivityManager.xaml.cs">
       <DependentUpon>ActivityManager.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Activities\Controls\SoundSlider.xaml.cs">
+      <DependentUpon>SoundSlider.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Activities\Controls\TrackPlayer.xaml.cs">
+      <DependentUpon>TrackPlayer.xaml</DependentUpon>
+    </Compile>
     <Compile Include="Activities\LibraryActivity.xaml.cs">
       <DependentUpon>LibraryActivity.xaml</DependentUpon>
     </Compile>
@@ -74,9 +80,6 @@
     </Compile>
     <Compile Include="Activities\Panels\ScreenPanel.xaml.cs">
       <DependentUpon>ScreenPanel.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Activities\Panels\SoundSliderPanel.xaml.cs">
-      <DependentUpon>SoundSliderPanel.xaml</DependentUpon>
     </Compile>
     <Compile Include="Activities\Panels\TagPanel.xaml.cs">
       <DependentUpon>TagPanel.xaml</DependentUpon>
@@ -107,6 +110,14 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
+    <Page Include="Activities\Controls\SoundSlider.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Activities\Controls\TrackPlayer.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
     <Page Include="Activities\LibraryActivity.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -120,10 +131,6 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Activities\Panels\ScreenPanel.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Activities\Panels\SoundSliderPanel.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
